重写重新生成的逻辑

+++ api/store/model/chat_message.go
@@ -5,7 +5,7 @@ import (
 )
 
 type ChatMessage struct {
-	Id          int64     `gorm:"column:id;primaryKey;autoIncrement" json:"id"`
+	Id          uint      `gorm:"column:id;primaryKey;autoIncrement" json:"id"`
 	UserId      uint      `gorm:"column:user_id;type:int;not null;comment:用户 ID" json:"user_id"`
 	ChatId      string    `gorm:"column:chat_id;type:char(40);not null;index;comment:会话 ID" json:"chat_id"`
 	Type        string    `gorm:"column:type;type:varchar(10);not null;comment:类型：prompt|reply" json:"type"`
